refactor : beverage 패키지 생성
Americano, Latte 클래스를 해당 패키지로 이동
